subjects: add to_anki for exporting subject to Anki
This required a refactor and is bundled with various touch-ups and
fixes.
wktoolkit was renamed to get_subjects. Args handling in get_subjects
is now scoped into its own function. A mutually-exclusive group of
arguments is added to support printing original JSON, an Anki format,
or the Anki schema. Members in Subjects and derived classes are
reordered as their order of definition is the Anki schema. A
Subject's aux_meaning is fixed to no longer incorrectly report the
primary meaning.
to_anki() is not fully implemented for Kanji and Vocabulary.